requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
python-jose[cryptography]==3.4.0
bcrypt==4.2.1
python-dotenv==1.1.0
email-validator==2.2.0
pytest==8.3.5
httpx==0.28.1
